tcp_conn_validator: Initial commit
Initial commit of the new provider tcp_conn_validator that allows
one to check if a service is actually listening on a given pair ip:host.
It supports both IPv4 and IPv6 address, it also support hostname.

It either takes a string connection in the namevar or received the
informations through the `host` and `port` parameters.

Case 1:

   tcp_conn_validator { '127.0.0.1:22' : }

Case 2:

   tcp_conn_validator { 'local ssh' :
     host => 'localhost',
     port => 22,
   }