tcp_conn_validator: Introduce the reachable param
This commit introduces the reachable param. It is by default set to
true, so any tcp_conn_validator will make sure the specified server is
listening on the specified port. When set to false, the check ensure
that the server is not listening to the specified ip:port pair.

It can allow one to take an action once a service has been really shut
down.